@@ -3,27 +3,25 @@
   "version": "0.1.0",
   "proxy": "http://127.0.0.1:5000",
   "private": true,
+  "engines": {
+    "node": ">=16.0.0"
+  },
   "dependencies": {
-    "@chakra-ui/icons": "^1.0.13",
-    "@chakra-ui/react": "^1.6.2",
-    "@emotion/react": "^11",
-    "@emotion/styled": "^11",
-    "@testing-library/jest-dom": "^5.11.4",
-    "@testing-library/react": "^11.1.0",
-    "@testing-library/user-event": "^12.1.10",
-    "axios": "^0.21.1",
-    "emoji-picker-react": "^4.4.9",
-    "framer-motion": "^4",
-    "react": "^16.14.0",
-    "react-chips": "^0.8.0",
-    "react-dom": "^16.14.0",
-    "react-lottie": "^1.2.3",
-    "react-notification-badge": "^1.5.1",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "4.0.3",
-    "react-scrollable-feed": "^1.3.1",
-    "socket.io-client": "^4.1.2",
-    "web-vitals": "^1.0.1"
+    "@chakra-ui/icons": "^2.1.1",
+    "@chakra-ui/react": "^2.8.2",
+    "@emotion/react": "^11.11.3",
+    "@emotion/styled": "^11.11.0",
+    "axios": "^1.6.7",
+    "emoji-picker-react": "^4.7.12",
+    "framer-motion": "^11.0.3",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-lottie": "^1.2.4",
+    "react-router-dom": "^6.22.0",
+    "react-scripts": "^5.0.1",
+    "react-scrollable-feed": "^2.0.1",
+    "socket.io-client": "^4.7.4",
+    "util": "^0.12.5"
   },
   "scripts": {
     "start": "react-scripts start",
