@@ -1,5 +1,5 @@
 const mongoose = require("mongoose");
-const colors = require("colors");
+// const colors = require("colors");
 
 const connectDB = async () => {
   try {
@@ -8,9 +8,9 @@ const connectDB = async () => {
       useUnifiedTopology: true,
     });
 
-    console.log(`MongoDB Connected: ${conn.connection.host}`.cyan.underline);
+    console.log(`MongoDB Connected: ${conn.connection.host}`);
   } catch (error) {
-    console.log(`Error: ${error.message}`.red.bold);
+    console.log(`Error: ${error.message}`);
     process.exit();
   }
 };
